requirements.txt
# Core dependencies
# requests==2.32.3
# beautifulsoup4==4.13.3
# pandas==2.2.3
# tqdm==4.67.1
# urllib3==2.3.0
# pyyaml==6.0.2
# spacy==3.8.4
# tiktoken==0.8.0
python-dotenv==1.0.1
ollama==0.4.7
langchain-core>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.24
streamlit>=1.31.0
elevenlabs>=0.3.0
langchain-ollama>=0.2.3
